+++ packages/listen/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0-beta.3
+
+- Updates `ErrorCallback` to pass the raw error object and original `StackTrace`.
+- Adds an optional `ErrorContext` parameter to `ErrorCallback` to distinguish listener errors from assertion errors.
+
 ## 1.0.0-beta.2
 
 - Fixes dart doc example directive.